rename gcloud.py to filesystem.py since it handles local ops also
