+++ fv3config/fv3run/_native.py
@@ -7,7 +7,7 @@ import tempfile
 import warnings
 import yaml
 from .._config import write_run_directory, _get_n_processes, _write_config_dict
-from .. import gcloud
+from .. import filesystem
 
 STDOUT_FILENAME = 'stdout.log'
 STDERR_FILENAME = 'stderr.log'
@@ -40,7 +40,10 @@ def run_native(config_dict_or_location, outdir, runfile=None):
             config_dict_or_location, config_out_filename)
         write_run_directory(config_dict, localdir)
         if runfile is not None:
-            gcloud._copy_file(runfile, os.path.join(localdir, os.path.basename(runfile)))
+            filesystem._get_file(
+                runfile,
+                os.path.join(localdir, os.path.basename(runfile))
+            )
         with _log_exceptions(localdir):
             n_processes = _get_n_processes(config_dict)
             _run_experiment(
@@ -66,9 +69,9 @@ def _temporary_directory(outdir):
             yield tempdir
         finally:
             logger.info('Copying output to %s', outdir)
-            fs = gcloud._get_fs(outdir)
+            fs = filesystem._get_fs(outdir)
             fs.makedirs(outdir, exist_ok=True)
-            gcloud._put_directory(tempdir, outdir)
+            filesystem._put_directory(tempdir, outdir)
 
 
 @contextlib.contextmanager
@@ -120,9 +123,9 @@ def _get_config_dict_and_write(config_dict_or_location, config_out_filename):
 
 
 def _copy_and_load_config_dict(config_location, local_target_location):
-    fs = gcloud._get_fs(config_location)
+    fs = filesystem._get_fs(config_location)
     fs.get(config_location, local_target_location)
-    gcloud._copy_file(config_location, local_target_location)
+    filesystem._copy_file(config_location, local_target_location)
     with open(local_target_location, 'r') as infile:
         config_dict = yaml.load(infile.read(), Loader=yaml.SafeLoader)
     return config_dict
